requirements.txt
asgiref==3.8.1
Django==5.1.8
django-debug-toolbar==5.1.0
django-environ==0.12.0
django-filter>=24.1
djangorestframework>=3.15.1
djangorestframework-simplejwt==5.5.0
faker==37.1.0
psycopg2-binary>=2.9.9
python-dotenv==1.0.1
sqlparse==0.5.3
